move materials for am in fe problem folder for external usage
